\name{ChIPprofileSimilarity-package}
\alias{ChIPprofileSimilarity-package}
\alias{ChIPprofileSimilarity}
\docType{package}
\title{
This package calculates metrics to estimate a level of similarity between two ChIP profiles.
}
\description{
TODO
}
\details{
\tabular{ll}{
Package: \tab ChIPprofileSimilarity\cr
Type: \tab Package\cr
Version: \tab 0.99\cr
Date: \tab 2014-08-08\cr
License: \tab Artistic-2.0\cr
}
}
\author{
Astrid Louise Deschenes <Astrid-Louise.Deschenes@crchuq.ulaval.ca>), 
Elsa Bertnachez <elsa.bernatchez.1@ulaval.ca>,
Charles Joly Beauparlant <charles.joly-beauparlant@crchul.ulaval.ca>,
Fabien Claude Lamaze <fabien.lamaze.1@ulaval.ca>,
Rawane Samb <rawane.samb.1@ulaval.ca>,
Pascal Belleau <Pascal.Belleau@crchudequebec.ulaval.ca> and
Arnaud Droit <arnaud.droit@crchuq.ulaval.ca>

Maintainer: Astrid Louise Deschenes <Astrid-Louise.Deschenes@crchuq.ulaval.ca>
}
\references{
TODO
\keyword{ package }
\seealso{
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
TODO
}
