% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\docType{data}
\name{SpearmanCorrelation}
\alias{SpearmanCorrelation}
\title{SpearmanCorrelation class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SpearmanCorrelation
}
\value{
The \code{SpearmanCorrelation$new} function returns a 
\code{SpearmanCorrelation} object which contains the information about 
the two profiles. It can be used, as many times 
needed, to calculate the specified metric.
}
\description{
An object which is a interface to calculate the Spearman's  

The \code{SpearmanCorrelation} object is needed to 
calculate the ratio between 
the peaks values between two profiles.
A threshold and the two profiles are set during the \code{SpearmanCorrelation} 
object creation. If different profiles are needed, the 
\code{calculateMetric} function should be used, 
with the new profiles passed as arguments to update those
values inside the \code{SpearmanCorrelation} object.
}
\section{Constructor}{

Create a \code{SpearmanCorrelation} object.

\code{SpearmanCorrelation$new(profile1, profile2, threshold = NULL)}
    
The \code{SpearmanCorrelation} object inherites those functions:
\itemize{
\item \code{getMetric} { A function that returns the value of the
calculated metric }
\item \code{getInfo} { A function that returns a description of the metric
with the metric value.}
\item \code{getType} { A function that returns the unique name associated
to this metric }
\item \code{calculateMetric} { A function that modifies the values of the
two profiles and the threshold. The new values (profile1, profile2) 
are passed as arguments.}
}
}
\author{
Astrid Deschenes
}
\seealso{
\itemize{
\item \code{\link{MetricFactory}} {for using a interface to calculate all 
available metrics separately or togheter.}
}
}
\keyword{datasets}

