\docType{package}
\name{ChIPprofileSimilarity}
\alias{ChIPprofileSimilarity-package}
\alias{ChIPprofileSimilarity}
\title{
ChIPprofileSimilarity: Metrics to estimate a level of similarity between two ChIP profiles
}
\description{
\pkg{ChIPprofileSimilarity} is calculing four differents metrics to estimate a level of similarity between two ChIP profiles.
}
\details{
The R function \code{\link{similarity}} calculates four differents metrics: the ratio of the maximum values, the ratio of the areas, the ratio between the intersection area and the total area, the difference between two profiles maximal peaks positions.

The function \code{\link{similarity}} also reports basic information about each ChIP profile such as the number of positions, the area, the maximum value and the position of the maximum value.

To learn more about \pkg{ChIPprofileSimilarity} package see:

\url{https://github.com/adeschen/ChIPprofileSimilarity/wiki}
}
\author{
A. Deschenes, 
E. Bertnachez,
C. Joly Beauparlant,
F.C. Lamaze,
R. Samb,
P. Belleau and
A. Droit 

Maintainer: Astrid Louise Deschenes <Astrid-Louise.Deschenes@crchuq.ulaval.ca>
}
\seealso{
\code{\link{similarity}}
}
\keyword{package}
